@@ -1,13 +1,12 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.1"
+    compileSdkVersion 28
 
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 22
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
 
@@ -23,9 +22,9 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.github.NodeMedia:NodeMediaClient-Android:2.4.2'
-    compile 'com.android.support:appcompat-v7:23.0.1'
-    compile "com.facebook.react:react-native:+"  // From node_modules
-	testCompile 'junit:junit:4.12'
+	implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'com.github.NodeMedia:NodeMediaClient-Android:2.6.0'
+    implementation 'com.android.support:appcompat-v7:28.0.1'
+    implementation "com.facebook.react:react-native:+"  // From node_modules
+	testImplementation 'junit:junit:4.12'
 }
