@@ -9,8 +9,8 @@ package cn.nodemedia.react_native_nodemediaclient;
 
 import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.common.MapBuilder;
-import com.facebook.react.uimanager.SimpleViewManager;
 import com.facebook.react.uimanager.ThemedReactContext;
+import com.facebook.react.uimanager.ViewGroupManager;
 import com.facebook.react.uimanager.annotations.ReactProp;
 
 import java.util.Map;
@@ -19,7 +19,7 @@ import javax.annotation.Nullable;
 
 import cn.nodemedia.react_native_nodemediaclient.RCTNodePlayerView;
 
-public class NodePlayerViewManager extends SimpleViewManager<RCTNodePlayerView> {
+public class NodePlayerViewManager extends ViewGroupManager<RCTNodePlayerView> {
 
 
     private static final int COMMAND_PAUSE_ID = 1;
